upgraded: Split daemon into multiple files separated by upgrade cause
Make the daemon easier to read by separating stream and node upgrades.
