+++ pkg/upgraded/daemon/daemon_test.go
@@ -9,14 +9,9 @@ import (
 	"time"
 
 	fleetlockclient "github.com/heathcliff26/fleetlock/pkg/server/client"
-	"github.com/heathcliff26/kube-upgrade/pkg/constants"
 	"github.com/heathcliff26/kube-upgrade/pkg/upgraded/config"
 	"github.com/heathcliff26/kube-upgrade/pkg/upgraded/fleetlock"
-	rpmostree "github.com/heathcliff26/kube-upgrade/pkg/upgraded/rpm-ostree"
 	"github.com/stretchr/testify/assert"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes/fake"
 )
 
 func TestNewDaemon(t *testing.T) {
@@ -168,131 +163,6 @@ func TestRetry(t *testing.T) {
 	t.Cleanup(cancel)
 
 	assert.Equal(t, 6, count, "Should have run the function exactly 6 times")
-}
-
-func TestDoUpgrade(t *testing.T) {
-	t.Run("LockAlreadyReserved", func(t *testing.T) {
-		assert := assert.New(t)
-
-		client, srv := NewFakeFleetlockServer(t, http.StatusLocked)
-		t.Cleanup(func() {
-			srv.Close()
-		})
-		rpmOstreeCMD, err := rpmostree.New("testdata/exit-0.sh")
-		if !assert.NoError(err, "Failed to create rpm-ostree command") {
-			t.FailNow()
-		}
-
-		d := &daemon{
-			fleetlock: client,
-			rpmostree: rpmOstreeCMD,
-		}
-
-		err = d.doUpgrade()
-
-		assert.ErrorContains(err, "failed to aquire lock:")
-	})
-	t.Run("FailedOstreeUpgrade", func(t *testing.T) {
-		assert := assert.New(t)
-
-		client, srv := NewFakeFleetlockServer(t, http.StatusOK)
-		t.Cleanup(func() {
-			srv.Close()
-		})
-		rpmOstreeCMD, err := rpmostree.New("testdata/exit-1.sh")
-		if !assert.NoError(err, "Failed to create rpm-ostree command") {
-			t.FailNow()
-		}
-
-		d := &daemon{
-			fleetlock: client,
-			rpmostree: rpmOstreeCMD,
-		}
-
-		err = d.doUpgrade()
-
-		assert.Error(err, "Should exit with error")
-	})
-	// This case is kinda scetchy, as in reality the system would reboot on success, thus the method should never return
-	t.Run("Success", func(t *testing.T) {
-		assert := assert.New(t)
-
-		client, srv := NewFakeFleetlockServer(t, http.StatusOK)
-		t.Cleanup(func() {
-			srv.Close()
-		})
-		rpmOstreeCMD, err := rpmostree.New("testdata/exit-0.sh")
-		if !assert.NoError(err, "Failed to create rpm-ostree command") {
-			t.FailNow()
-		}
-
-		d := &daemon{
-			fleetlock: client,
-			rpmostree: rpmOstreeCMD,
-		}
-
-		err = d.doUpgrade()
-
-		assert.NoError(err, "Should succeed")
-	})
-}
-
-func TestUpdateNodeStatus(t *testing.T) {
-	tMatrix := []struct {
-		Name  string
-		Node  *corev1.Node
-		Error bool
-	}{
-		{
-			Name: "Success",
-			Node: &corev1.Node{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "testnode",
-					Annotations: map[string]string{
-						constants.KubernetesUpgradeStatus: "unset",
-					},
-				},
-			},
-		},
-		{
-			Name: "NoAnnotations",
-			Node: &corev1.Node{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "testnode",
-				},
-			},
-		},
-		{
-			Name:  "NoNode",
-			Error: true,
-		},
-	}
-
-	for _, tCase := range tMatrix {
-		t.Run(tCase.Name, func(t *testing.T) {
-			assert := assert.New(t)
-
-			c := fake.NewSimpleClientset()
-			d := &daemon{
-				client: c,
-				ctx:    context.Background(),
-			}
-			if tCase.Node != nil {
-				_, _ = c.CoreV1().Nodes().Create(context.Background(), tCase.Node, metav1.CreateOptions{})
-				d.node = tCase.Node.GetName()
-			} else {
-				d.node = "not-a-node"
-			}
-
-			if tCase.Error {
-				assert.Error(d.updateNodeStatus("new-status"), "Should fail")
-			} else {
-				assert.NoError(d.updateNodeStatus("new-status"), "Should succeed")
-				node, _ := c.CoreV1().Nodes().Get(context.Background(), d.node, metav1.GetOptions{})
-				assert.Equal("new-status", node.GetAnnotations()[constants.KubernetesUpgradeStatus], "Should have set status")
-			}
-		})
-	}
 }
 
 func cancelOnTimeout(t *testing.T, ctx context.Context, cancel context.CancelFunc) {
